creating pages home, about and login + implementing router logic
