+++ reactApp/src/components/authentication/Login.js
@@ -1,0 +1,9 @@
+function Login() {
+  return (
+    <>
+      <h1>(MODAL) Login with your Naturdoc Account </h1>
+    </>
+  );
+}
+
+export default Login;
